feat: calculate factorial (with some difficult)

+++ factorial-worker/src/main/java/com/calcltda/factorialworker/factorial/FactorialDTO.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public record FactorialDTO(
         long id,
-        long number
+        int number
 ) implements Serializable {
 
 }
